Add `--deploy-key` parameter
This adds a `--deploy-key` parameter for syncing the deploy keys of a
Github repositories.

You can use it in combination with the `--user` and/or `--organization`
parameters.

To be able to access deploy keys through Githubs API requires an access
token with the `public_repo` or `repo` (for private repository) scope.

The test case is in its own file because of the dependency on a token
use the build tag to test: `go test -tags=deploy_key ./...`.